requirements.txt
fastapi
firebase-admin
google-cloud-storage
mangum
pydantic
python-dotenv
sendgrid
stripe
uvicorn
boto3